refactor: ドメイン例外を共有カーネルに移動
汎用的な例外クラスを共有カーネルに移動し、他のモジュールでも再利用できるようにしました。

移動した例外:
- DomainException (基底クラス)
- NotFoundException
- ValidationException
- BusinessRuleException
- およびその派生クラス

食材管理固有の例外は引き続きモジュール内に保持:
- CategoryNotFoundException
- UnitNotFoundException
