+++ src/modules/ingredients/server/api/handlers/commands/create-ingredient.handler.ts
@@ -3,7 +3,7 @@ import { ZodError } from 'zod'
 import { CreateIngredientCommand } from '../../../application/commands/create-ingredient.command'
 import { CreateIngredientHandler } from '../../../application/commands/create-ingredient.handler'
 import { IngredientMapper } from '../../../application/mappers/ingredient.mapper'
-import { ValidationException } from '../../../domain/exceptions/validation.exception'
+import { ValidationException } from '../../../domain/exceptions'
 import { CategoryRepository } from '../../../domain/repositories/category-repository.interface'
 import { UnitRepository } from '../../../domain/repositories/unit-repository.interface'
 import { CategoryId, StorageType, UnitId } from '../../../domain/value-objects'
