+++ src/modules/ingredients/server/domain/exceptions/not-found.exception.ts
@@ -1,20 +1,4 @@
-import { DomainException } from './domain.exception'
-
-/**
- * リソースが見つからない場合の例外
- */
-export class NotFoundException extends DomainException {
-  readonly httpStatusCode = 404
-  readonly errorCode = 'RESOURCE_NOT_FOUND'
-
-  constructor(resourceType: string, identifier: string | Record<string, unknown>) {
-    const id = typeof identifier === 'string' ? identifier : JSON.stringify(identifier)
-    super(`${resourceType} not found: ${id}`, {
-      resourceType,
-      identifier,
-    })
-  }
-}
+import { NotFoundException } from '@/modules/shared/server/domain/exceptions'
 
 /**
  * カテゴリが見つからない場合の例外
