+++ src/modules/ingredients/server/domain/value-objects/ingredient-id.vo.ts
@@ -2,7 +2,7 @@ import { randomUUID } from 'crypto'
 
 import { ValueObject } from '@/modules/shared/server/domain/value-objects/value-object.base'
 
-import { ValidationException } from '../exceptions/validation.exception'
+import { ValidationException } from '../exceptions'
 
 /**
  * 食材ID値オブジェクト
