+++ src/modules/shared/server/domain/exceptions/business-rule.exception.ts
