+++ src/modules/shared/server/domain/exceptions/domain.exception.ts
