+++ src/modules/shared/server/domain/exceptions/index.ts
@@ -1,0 +1,19 @@
+// 基底クラス
+export { DomainException } from './domain.exception'
+
+// リソース不在例外
+export { NotFoundException } from './not-found.exception'
+
+// バリデーション例外
+export {
+  ValidationException,
+  RequiredFieldException,
+  InvalidFieldException,
+} from './validation.exception'
+
+// ビジネスルール例外
+export {
+  BusinessRuleException,
+  DuplicateException,
+  OperationNotAllowedException,
+} from './business-rule.exception'
