+++ src/modules/shared/server/domain/exceptions/validation.exception.ts
