+++ src/modules/shared/server/domain/index.ts
@@ -4,3 +4,4 @@
  */
 
 export * from './value-objects'
+export * from './exceptions'
