+++ tests/unit/modules/ingredients/server/domain/value-objects/expiry-info.vo.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect } from 'vitest'
 
-import { ValidationException } from '@/modules/ingredients/server/domain/exceptions/validation.exception'
+import { ValidationException } from '@/modules/ingredients/server/domain/exceptions'
 import { ExpiryInfo } from '@/modules/ingredients/server/domain/value-objects'
 
 import { ExpiryInfoBuilder } from '../../../../../../__fixtures__/builders'
